Fixing build warnings, corrected mod directory loc.

+++ EU4toV2/Source/EU4World/Mods/Mods.cpp
@@ -227,7 +227,7 @@ bool EU4::Mods::extractZip(const std::string& archive, const std::string& path) 
 		return false;
 	for (size_t entryNum = 0; entryNum < modfile->GetEntriesCount(); ++entryNum)
 	{
-		const auto& entry = modfile->GetEntry(entryNum);
+		const auto& entry = modfile->GetEntry(static_cast<int>(entryNum));
 		const auto& inpath = entry->GetFullName();
 		const auto& name = entry->GetName();
 		if (entry->IsDirectory())
